SOF-1420 Add GfxShowMapping to Qbox. Reduce duplication of configuration
