+++ src/tv2_offtube_showstyle/__tests__/config-manifest.spec.ts
@@ -16,6 +16,7 @@ const blankShowStyleConfig: OfftubeShowStyleConfig = {
 	MakeAdlibsForFulls: true,
 	GfxSchemaTemplates: [],
 	GfxSetups: [],
+	GfxShowMapping: [],
 	GfxDefaults: []
 }
